Extract dashboard helper functions
